requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.41
psycopg[binary]==3.2.9
psycopg2-binary==2.9.10
python-dotenv==1.0.0
cryptography==41.0.7
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.12.1
pydantic-settings==2.10.1
requests==2.32.3
python-multipart==0.0.20
supabase

